+++ lib/c-api/src/wasm_c_api/module.rs
@@ -41,16 +41,17 @@ pub struct wasm_module_t {
 ///     // Create a WebAssembly module from a WAT definition.
 ///     wasm_byte_vec_t wat;
 ///     wasmer_byte_vec_new_from_string(&wat, "(module)");
-///     wasm_byte_vec_t* wasm = wat2wasm(&wat);
-///
+///     wasm_byte_vec_t wasm;
+///     wat2wasm(&wat, &wasm);
+///    
 ///     // Create the module.
-///     wasm_module_t* module = wasm_module_new(store, wasm);
-///
+///     wasm_module_t* module = wasm_module_new(store, &wasm);
+
 ///     // It works!
 ///     assert(module);
-///
+///    
 ///     // Free everything.
-///     wasm_byte_vec_delete(wasm);
+///     wasm_byte_vec_delete(&wasm);
 ///     wasm_byte_vec_delete(&wat);
 ///     wasm_module_delete(module);
 ///     wasm_store_delete(store);
@@ -109,13 +110,14 @@ pub unsafe extern "C" fn wasm_module_delete(_module: Option<Box<wasm_module_t>>)
 ///     // Create a WebAssembly module from a WAT definition.
 ///     wasm_byte_vec_t wat;
 ///     wasmer_byte_vec_new_from_string(&wat, "(module)");
-///     wasm_byte_vec_t* wasm = wat2wasm(&wat);
+///     wasm_byte_vec_t wasm;
+///     wat2wasm(&wat, &wasm);
 ///
 ///     // Validate that the WebAssembly module is correct.
 ///     assert(wasm_module_validate(store, wasm));
 ///
 ///     // Free everything.
-///     wasm_byte_vec_delete(wasm);
+///     wasm_byte_vec_delete(&wasm);
 ///     wasm_byte_vec_delete(&wat);
 ///     wasm_store_delete(store);
 ///     wasm_engine_delete(engine);
@@ -182,7 +184,8 @@ pub unsafe extern "C" fn wasm_module_validate(
 ///         "  (table (export \"table\") 0 funcref)\n"
 ///         "  (memory (export \"memory\") 1))"
 ///     );
-///     wasm_byte_vec_t* wasm = wat2wasm(&wat);
+///     wasm_byte_vec_t wasm;
+///     wat2wasm(&wat, &wasm);
 ///
 ///     // Create the module.
 ///     wasm_module_t* module = wasm_module_new(store, wasm);
@@ -249,7 +252,7 @@ pub unsafe extern "C" fn wasm_module_validate(
 ///
 ///     // Free everything.
 ///     wasm_exporttype_vec_delete(&export_types);
-///     wasm_byte_vec_delete(wasm);
+///     wasm_byte_vec_delete(&wasm);
 ///     wasm_byte_vec_delete(&wat);
 ///     wasm_module_delete(module);
 ///     wasm_store_delete(store);
@@ -305,7 +308,8 @@ pub unsafe extern "C" fn wasm_module_exports(
 ///         "  (import \"ns\" \"table\" (table 1 2 anyfunc))\n"
 ///         "  (import \"ns\" \"memory\" (memory 3 4)))"
 ///     );
-///     wasm_byte_vec_t* wasm = wat2wasm(&wat);
+///     wasm_byte_vec_t wasm;
+///     wat2wasm(&wat, &wasm);
 ///
 ///     // Create the module.
 ///     wasm_module_t* module = wasm_module_new(store, wasm);
@@ -390,7 +394,7 @@ pub unsafe extern "C" fn wasm_module_exports(
 ///     // Free everything.
 ///     wasm_importtype_vec_delete(&import_types);
 ///     wasm_module_delete(module);
-///     wasm_byte_vec_delete(wasm);
+///     wasm_byte_vec_delete(&wasm);
 ///     wasm_byte_vec_delete(&wat);
 ///     wasm_store_delete(store);
 ///     wasm_engine_delete(engine);
@@ -456,7 +460,8 @@ pub unsafe extern "C" fn wasm_module_imports(
 ///        "  (table (export \"table\") 0 funcref)\n"
 ///        "  (memory (export \"memory\") 1))"
 ///    );
-///    wasm_byte_vec_t* wasm = wat2wasm(&wat);
+///    wasm_byte_vec_t wasm;
+///    wat2wasm(&wat, &wasm);
 ///
 ///    // Create the module.
 ///    wasm_module_t* module = wasm_module_new(store, wasm);
@@ -488,7 +493,7 @@ pub unsafe extern "C" fn wasm_module_imports(
 ///    // Free everything.
 ///    wasm_exporttype_vec_delete(&export_types);
 ///    wasm_module_delete(deserialized_module);
-///    wasm_byte_vec_delete(wasm);
+///    wasm_byte_vec_delete(&wasm);
 ///    wasm_byte_vec_delete(&wat);
 ///    wasm_store_delete(store);
 ///    wasm_engine_delete(engine);
@@ -561,11 +566,12 @@ mod tests {
 
                 wasm_byte_vec_t wat;
                 wasmer_byte_vec_new_from_string(&wat, "(module)");
-                wasm_byte_vec_t* wasm = wat2wasm(&wat);
+                wasm_byte_vec_t wasm;
+                wat2wasm(&wat, &wasm);
 
-                assert(wasm_module_validate(store, wasm));
+                assert(wasm_module_validate(store, &wasm));
 
-                wasm_byte_vec_delete(wasm);
+                wasm_byte_vec_delete(&wasm);
                 wasm_byte_vec_delete(&wat);
                 wasm_store_delete(store);
                 wasm_engine_delete(engine);
@@ -587,12 +593,13 @@ mod tests {
 
                 wasm_byte_vec_t wat;
                 wasmer_byte_vec_new_from_string(&wat, "(module)");
-                wasm_byte_vec_t* wasm = wat2wasm(&wat);
+                wasm_byte_vec_t wasm;
+                wat2wasm(&wat, &wasm);
 
-                wasm_module_t* module = wasm_module_new(store, wasm);
+                wasm_module_t* module = wasm_module_new(store, &wasm);
                 assert(module);
 
-                wasm_byte_vec_delete(wasm);
+                wasm_byte_vec_delete(&wasm);
                 wasm_byte_vec_delete(&wat);
                 wasm_module_delete(module);
                 wasm_store_delete(store);
@@ -622,9 +629,10 @@ mod tests {
                     "  (table (export \"table\") 0 funcref)\n"
                     "  (memory (export \"memory\") 1))"
                 );
-                wasm_byte_vec_t* wasm = wat2wasm(&wat);
+                wasm_byte_vec_t wasm;
+                wat2wasm(&wat, &wasm);
 
-                wasm_module_t* module = wasm_module_new(store, wasm);
+                wasm_module_t* module = wasm_module_new(store, &wasm);
                 assert(module);
 
                 wasm_exporttype_vec_t export_types;
@@ -699,7 +707,7 @@ mod tests {
                 }
 
                 wasm_exporttype_vec_delete(&export_types);
-                wasm_byte_vec_delete(wasm);
+                wasm_byte_vec_delete(&wasm);
                 wasm_byte_vec_delete(&wat);
                 wasm_module_delete(module);
                 wasm_store_delete(store);
@@ -729,9 +737,10 @@ mod tests {
                     "  (import \"ns\" \"table\" (table 1 2 anyfunc))\n"
                     "  (import \"ns\" \"memory\" (memory 3 4)))"
                 );
-                wasm_byte_vec_t* wasm = wat2wasm(&wat);
+                wasm_byte_vec_t wasm;
+                wat2wasm(&wat, &wasm);
 
-                wasm_module_t* module = wasm_module_new(store, wasm);
+                wasm_module_t* module = wasm_module_new(store, &wasm);
                 assert(module);
 
                 wasm_importtype_vec_t import_types;
@@ -817,7 +826,7 @@ mod tests {
 
                 wasm_importtype_vec_delete(&import_types);
                 wasm_module_delete(module);
-                wasm_byte_vec_delete(wasm);
+                wasm_byte_vec_delete(&wasm);
                 wasm_byte_vec_delete(&wat);
                 wasm_store_delete(store);
                 wasm_engine_delete(engine);
@@ -839,9 +848,10 @@ mod tests {
 
                 wasm_byte_vec_t wat;
                 wasmer_byte_vec_new_from_string(&wat, "(module)");
-                wasm_byte_vec_t* wasm = wat2wasm(&wat);
+                wasm_byte_vec_t wasm;
+                wat2wasm(&wat, &wasm);
 
-                wasm_module_t* module = wasm_module_new(store, wasm);
+                wasm_module_t* module = wasm_module_new(store, &wasm);
                 assert(module);
 
                 wasm_byte_vec_t serialized_module;
@@ -850,7 +860,7 @@ mod tests {
 
                 wasm_module_delete(module);
                 wasm_byte_vec_delete(&serialized_module);
-                wasm_byte_vec_delete(wasm);
+                wasm_byte_vec_delete(&wasm);
                 wasm_byte_vec_delete(&wat);
                 wasm_store_delete(store);
                 wasm_engine_delete(engine);
@@ -879,9 +889,10 @@ mod tests {
                     "  (table (export \"table\") 0 funcref)\n"
                     "  (memory (export \"memory\") 1))"
                 );
-                wasm_byte_vec_t* wasm = wat2wasm(&wat);
+                wasm_byte_vec_t wasm;
+                wat2wasm(&wat, &wasm);
 
-                wasm_module_t* module = wasm_module_new(store, wasm);
+                wasm_module_t* module = wasm_module_new(store, &wasm);
                 assert(module);
 
                 wasm_byte_vec_t serialized_module;
@@ -903,7 +914,7 @@ mod tests {
 
                 wasm_exporttype_vec_delete(&export_types);
                 wasm_module_delete(deserialized_module);
-                wasm_byte_vec_delete(wasm);
+                wasm_byte_vec_delete(&wasm);
                 wasm_byte_vec_delete(&wat);
                 wasm_store_delete(store);
                 wasm_engine_delete(engine);