+++ lib/compiler/src/lib.rs
@@ -86,8 +86,8 @@ pub use crate::target::{
 #[cfg(feature = "translator")]
 pub use crate::translator::{
     to_wasm_error, translate_module, wptype_to_type, FunctionBodyData, FunctionMiddleware,
-    FunctionMiddlewareGenerator, GenerateMiddlewareChain, MiddlewareBinaryReader,
-    MiddlewareReaderState, ModuleEnvironment, ModuleInfoTranslation, ModuleTranslationState,
+    MiddlewareBinaryReader, MiddlewareReaderState, ModuleEnvironment, ModuleInfoTranslation,
+    ModuleMiddleware, ModuleMiddlewareChain, ModuleTranslationState,
 };
 pub use crate::trap::TrapInformation;
 pub use crate::unwind::CompiledFunctionUnwindInfo;